requirements.txt
# Core Dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
python-multipart>=0.0.6

# MCP Protocol
websockets>=12.0
jsonrpc-websocket>=3.1.0

# Database
chromadb>=0.4.20
sqlite-utils>=3.35.0
aiosqlite>=0.19.0  # Optional: Better async SQLite performance
sqlalchemy>=2.0.25

# Text Processing
pypdf2>=3.0.1
pdfplumber>=0.10.0
ebooklib>=0.18
spacy>=3.7.0
nltk>=3.8.0
python-magic>=0.4.27

# OCR Support
pytesseract>=0.3.10
pdf2image>=1.16.0
opencv-python>=4.9.0

# Embeddings
openai>=1.10.0
sentence-transformers>=2.2.0
torch>=2.5.0
transformers>=4.36.0

# Math Processing
sympy>=1.12
latex2sympy2>=1.9.0

# Code Processing
pygments>=2.17.0
black>=23.12.0

# C++ Bindings
pybind11>=2.11.0
cmake>=3.28.0

# Utilities
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.7.0
tqdm>=4.66.0
pyyaml>=6.0.0

# Caching
redis>=5.0.0
cachetools>=5.3.0

# For development dependencies, install with:
# pip install -r requirements-dev.txtaiosqlite>=0.20.0
bcrypt>=4.0.0
